requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
pandas==1.2.1
pynacl==1.4.0
cogwatch==2.0.0